renamed new feedback fiels in the exchange
