+++ nbgrader/exchange/__init__.py
@@ -1,10 +1,10 @@
 from .exchange import Exchange, ExchangeError
 from .collect import ExchangeCollect
 from .fetch import ExchangeFetch
-from .fetchfeedback import ExchangeFetchFeedback
+from .fetch_feedback import ExchangeFetchFeedback
 from .list import ExchangeList
 from .release import ExchangeRelease
-from .releaseFeedback import ExchangeReleaseFeedback
+from .release_feedback import ExchangeReleaseFeedback
 from .submit import ExchangeSubmit
 
 __all__ = [
